+++ src/ApiInterface.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Ifrost\ApiFoundation;
 
-use Exception;
 use Ifrost\ApiFoundation\Exception\NotFoundException;
 use Ifrost\ApiFoundation\Exception\NotUniqueException;
 use Symfony\Component\HttpFoundation\Response;
@@ -12,19 +11,19 @@ use Symfony\Component\HttpFoundation\Response;
 interface ApiInterface
 {
     /**
-     * @throws Exception
+     * @throws \Exception
      */
     public function find(): Response;
 
     /**
      * @throws NotFoundException
-     * @throws Exception
+     * @throws \Exception
      */
     public function findOne(): Response;
 
     /**
      * @throws NotUniqueException
-     * @throws Exception
+     * @throws \Exception
      */
     public function create(): Response;
 
@@ -32,7 +31,7 @@ interface ApiInterface
      * Replaces the entire object data with the given data.
      *
      * @throws NotFoundException
-     * @throws Exception
+     * @throws \Exception
      */
     public function update(): Response;
 
@@ -40,12 +39,12 @@ interface ApiInterface
      * Modify only selected fields.
      *
      * @throws NotFoundException
-     * @throws Exception
+     * @throws \Exception
      */
     public function modify(): Response;
 
     /**
-     * @throws Exception
+     * @throws \Exception
      */
     public function delete(): Response;
 }