+++ tests/Unit/Routing/AbstractApiLoader/SupportsTest.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace Ifrost\ApiFoundation\Tests\Unit\Routing\AbstractApiLoader;
 
-use PHPUnit\Framework\TestCase;
-use Symfony\Component\Config\FileLocator;
 use Ifrost\ApiFoundation\Tests\Variant\Routing\AnnotatedRouteControllerLoader;
 use Ifrost\ApiFoundation\Tests\Variant\Routing\ApiLoader;
+use PHPUnit\Framework\TestCase;
+use Symfony\Component\Config\FileLocator;
 
 class SupportsTest extends TestCase
 {