+++ tests/Variant/Controller/ProductControllerWithOverwrittenAction.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace Ifrost\ApiFoundation\Tests\Variant\Controller;
 
 use Ifrost\ApiFoundation\Attribute\Api;
+use Ifrost\ApiFoundation\Tests\Variant\Entity\Product;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
-use Ifrost\ApiFoundation\Tests\Variant\Entity\Product;
 
 #[Api(entity: Product::class, path: 'products')]
 class ProductControllerWithOverwrittenAction extends ApiControllerVariant